Move ObserveCenters into callbacks directory
