+++ src/ParallelAlgorithms/ApparentHorizonFinder/Callbacks/CMakeLists.txt
@@ -9,5 +9,6 @@ spectre_target_headers(
   ErrorOnFailedApparentHorizon.hpp
   FindApparentHorizon.hpp
   IgnoreFailedApparentHorizon.hpp
+  ObserveCenters.hpp
   SendDependencyToObserverWriter.hpp
   )
